Use slog.UInt64 instead of slog.Int with uint types

+++ server/handle_map_layer_zxy.go
@@ -230,9 +230,9 @@ func (req HandleMapLayerZXY) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		slog.Default().Info("tile is rather large", 
 			slog.String("map", req.mapName),
 			slog.String("layer", req.layerName),
-			slog.Int("z", int(req.z)),
-			slog.Int("x", int(req.x)),
-			slog.Int("y", int(req.y)),
+			slog.UInt64("z", req.z),
+			slog.UInt64("x", req.x),
+			slog.UInt64("y", req.y),
 			slog.Int("size_kb", len(pbyte)/1024),
 		)
 	}
